Crate xplm

The xplm crate provides a convenient interface to the X-Plane plugin SDK.

Modules

command
Functionality for finding, creating, and executing commands

data
Functionality for reading, writing, and creating datarefs

features
SDK feature querying and enabling

flight_loop
Flight loop callbacks

frequency
Radio frequency representation

graphics
OpenGL-related functionality

nav
Provides access to the navigation database

position
Position types

terrain
Terrain probing

ui
User interface elements

Functions

debug
Writes a message to the X-Plane log.txt file

enable_debug_logging
Enables the logging of plugin-related errors to the log.txt file

find_symbol
Finds a symbol in the set of currently loaded libraries