Struct xplm::command::Command

pub struct Command<O> {
    // some fields omitted
}

A command

The type parameter O indicates whether this plugin or something else owns the command.

A command execution has three phases:

The command begins. An external command is begun by calling Command::begin(). The callback's command_begin() method is called. This corresponds to a physical button being pressed down.

The callback continues. X-Plane calls the callback's command_continue method frequently until the command execution ends.

The command ends. An external command is ended by calling Command::end(). The callback's command_end() method is called. This corresponds to a physical button being released.

Owned commands

Owned commands are created by a plugin. A callback can be set, which will be called when something executes the command.

When an owned command goes out of scope, its callback also goes out of scope. However, because X-Plane does not allow commands to be deleted at runtime, the command will still exist. Excecuting the command will have not effect.

External commands

External commands are created by X-Plane or another plugin. Because the order of plugin loading is not specified and commands created by other plugins are not available until they load, plugins should not search for other plugins' commands until the plugin enable callback.

Examples

Finding and executing a command

let command = Command::find("sim/systems/avionics_on").unwrap();
command.begin();
command.end();

Creating a command

struct TestCommandCallback;

impl CommandCallback for TestCommandCallback {
    fn command_begin(&mut self) {
        println!("Command begun");
    }
    fn command_continue(&mut self) {
        println!("Command continues");
    }
    fn command_end(&mut self) {
        println!("Command ended");
    }
}

let mut command = Command::create("test/my_plugin/do_something", "Does the thing").unwrap();
command.set_callback(TestCommandCallback);

Methods

impl Command<External>

fn find(name: &str) -> Result<Command<External>, SearchError>

Finds a command by its name

fn begin(&self)

Starts executing this command

fn end(&self)

Finishes executing this command

impl Command<Owned>

fn create(name: &str, description: &str) -> Result<Command<Owned>, CreateError>

Creates a command

name: The name of the command

description: Brief text that describes the command

fn set_callback<C>(&mut self, callback: C) where C: 'static + CommandCallback

Sets the callback that will be called when the command is executed.

This Command object takes ownership of the callback.