Module xplm::data

Functionality for reading, writing, and creating datarefs

Structs

Borrowed
Provides safe access to a dataref that X-Plane or another plugin has already created

Owned
A dataref created and owned by this plugin

ReadOnly
Marks a dataref that can only be read

ReadWrite
Marks a dataref that can be read and written

Shared
Provides access to a shared dataref

Enums

SearchError
Possible errors encountered when finding a dataref

Traits

ArrayReadable
A trait for objects in which an array of values can be stored

ArrayWriteable
A trait for objects that can be written like arrays

DataAccess
Trait for a read/write or read-only marker

DataType
Trait for types that have associated type IDs in X-Plane

Readable
A trait for objects that can be read to get a value

StringReadable
A trait for objects that can be read as Strings

StringWriteable
A trait for objects that can be written as Strings

Writeable
A trait for objects in which a value can be stored