Struct xplm::data::dataref::DataRef

pub struct DataRef<D, A> {
    // some fields omitted
}

Provides safe access to a dataref

The type parameter D specifies the type of the dataref. The following types are supported:

i32: X-Plane type int

f32: X-Plane type float

f64: X-Plane type double

Vec<i32>: X-Plane type int array

Vec<f32>: X-Plane type float array

Vec<u8>: X-Plane type byte array

String: X-Plane type byte array

The type parameter A specifies the writeability of the dataref. This should be either ReadWrite for a writeable dataref or ReadOnly for a read-only dataref.

When a DataRef is created, its type and writeability are checked against the type and writeability specified by X-Plane. A DataRef can only be created if X-Plane knows about a dataref with the same name and compatible type and writeabiltiy.

Examples

Read-only f32 dataref

let time_ref: DataRef<f32, ReadOnly> = DataRef::find("sim/time/total_running_time_sec").unwrap();
let time = time_ref.get();

Read-only String dataref

let dataref: DataRef<String, ReadOnly> = DataRef::find("sim/version/sim_build_string").unwrap();
let sim_build_time = dataref.get();

Writeable i32 dataref

let dataref: DataRef<i32, ReadWrite> = DataRef::find("sim/cockpit2/autopilot/flight_director_mode").unwrap();
let mode = dataref.get();
dataref.set(3);

Methods

impl<D, A> DataRef<D, A> where D: DataType, A: DataAccess

fn find(name: &str) -> Result<DataRef<D, A>, SearchError>

Finds a dataref with the provided name. Returns a DataRef object or an error

impl<A> DataRef<i32, A> where A: DataAccess

fn get(&self) -> i32

Returns the value of this dataref

impl DataRef<i32, ReadWrite>

fn set(&mut self, value: i32)

Sets the value of this dataref

impl<A> DataRef<f32, A> where A: DataAccess

fn get(&self) -> f32

Returns the value of this dataref

impl DataRef<f32, ReadWrite>

fn set(&mut self, value: f32)

Sets the value of this dataref

impl<A> DataRef<f64, A> where A: DataAccess

fn get(&self) -> f64

Returns the value of this dataref

impl DataRef<f64, ReadWrite>

fn set(&mut self, value: f64)

Sets the value of this dataref

impl<A> DataRef<Vec<i32>, A> where A: DataAccess

fn get(&self, data: &mut [i32])

Replaces values in the provided slice with the values stored in this dataref

If the dataref has more values than the length of the slice, values beyond the last element of the slice will not be copied.

If the dataref has fewer values than the length of the slice, the content of entries in the slice after the last entry in the dataref is unspecified.

If the slice has more than i32::max_value() elements, no more than i32::max_value() elements will be read.

fn len(&self) -> usize

Returns the number of values in this dataref

impl DataRef<Vec<i32>, ReadWrite>

fn set(&mut self, data: &[i32])

Writes values from the provided slice into this dataref

If the dataref has more values than the length of the slice, values beyond the last element of the slice will not be modified.

If the slice has more than i32::max_value() elements, no more than i32::max_value() elements will be written.

impl<A> DataRef<Vec<f32>, A> where A: DataAccess

fn get(&self, data: &mut [f32])

Replaces values in the provided slice with the values stored in this dataref

If the dataref has more values than the length of the slice, values beyond the last element of the slice will not be copied.

If the dataref has fewer values than the length of the slice, the content of entries in the slice after the last entry in the dataref is unspecified.

If the slice has more than i32::max_value() elements, no more than i32::max_value() elements will be read.

fn len(&self) -> usize

Returns the number of values in this dataref

impl DataRef<Vec<f32>, ReadWrite>

fn set(&mut self, data: &[f32])

Writes values from the provided slice into this dataref

If the dataref has more values than the length of the slice, values beyond the last element of the slice will not be modified.

If the slice has more than i32::max_value() elements, no more than i32::max_value() elements will be written.

impl<A> DataRef<Vec<u8>, A> where A: DataAccess

fn get(&self, data: &mut [u8])

Replaces values in the provided slice with the values stored in this dataref

If the dataref has more values than the length of the slice, values beyond the last element of the slice will not be copied.

If the dataref has fewer values than the length of the slice, the content of entries in the slice after the last entry in the dataref is unspecified.

If the slice has more than i32::max_value() elements, no more than i32::max_value() elements will be read.

fn len(&self) -> usize

Returns the number of values in this dataref

impl DataRef<Vec<u8>, ReadWrite>

fn set(&mut self, data: &[u8])

Writes values from the provided slice into this dataref

If the dataref has more values than the length of the slice, values beyond the last element of the slice will not be modified.

If the slice has more than i32::max_value() elements, no more than i32::max_value() elements will be written.

impl<A> DataRef<String, A> where A: DataAccess

fn get(&self) -> String

Reads and returns the value of this dataref

impl DataRef<String, ReadWrite>

fn set(&mut self, value: &str) -> Result<(), NulError>

Sets the value of this dataref

If the provided value contains more than i32::max_value() bytes, up to i32::max_value() bytes will be written.