Struct xplm::data::owned_data::OwnedData

pub struct OwnedData<D, A> {
    // some fields omitted
}

Provides safe access to a dataref owned by this plugin

The type parameter D specifies the type of the dataref. The following types are supported:

i32: X-Plane type int

f32: X-Plane type float

f64: X-Plane type double

Vec<i32>: X-Plane type int array

Vec<f32>: X-Plane type float array

Vec<u8>: X-Plane type byte array

The type parameter A specifies the writeability of the dataref. This should be either ReadWrite for a writeable dataref or ReadOnly for a read-only dataref.

Methods

impl<D, A> OwnedData<D, A> where D: DataType, A: DataAccess

fn create(name: &str, initial_value: D) -> Result<OwnedData<D, A>, NulError>

Creates a dataref with the provided name, set to the provided value

Returns the dataref on success, or an error if the provided name was invalid.

fn get(&self) -> D

Returns the value of this dataref

fn set(&mut self, value: D)

Sets the value of this dataref

impl<E, A> OwnedData<Vec<E>, A> where E: Clone

fn get_as_slice<'a>(&'a self) -> &'a [E]

Returns the value of this dataref as a slice

fn get_as_slice_mut<'a>(&'a mut self) -> &'a mut [E]

Returns the value of this dataref as a mutable slice

fn set_as_slice(&mut self, values: &[E])

Sets the value of this dataref to equal the content of the provided slice

impl<A> OwnedData<Vec<u8>, A>

fn get_as_string(&self) -> String

Returns the value of this dataref as a String

fn set_as_string(&mut self, value: &str)

Sets this dataref to equal a string. If the provided string contains one or more null bytes, the dataref is not changed. The dataref will be set to the bytes in the provided string, with a terminating null byte.