Struct xplm::data::Borrowed

pub struct Borrowed<D, A> {
    // some fields omitted
}

Provides safe access to a dataref that X-Plane or another plugin has already created

The type parameter D specifies the type of the dataref. The following types are supported:

i32: X-Plane type int

f32: X-Plane type float

f64: X-Plane type double

Vec<i32>: X-Plane type int array

Vec<f32>: X-Plane type float array

Vec<u8>: X-Plane type byte array

String: X-Plane type byte array

The type parameter A specifies the writeability of the dataref. This should be either ReadWrite for a writeable dataref or ReadOnly for a read-only dataref.

When a Borrowed is created, its type and writeability are checked against the type and writeability specified by X-Plane. A Borrowed can only be created if X-Plane knows about a dataref with the same name and compatible type and writeabiltiy.

Examples

Read-only f32 dataref

let time_ref: Borrowed<f32, ReadOnly>
    = Borrowed::find("sim/time/total_running_time_sec").unwrap();
let time = time_ref.get();

Read-only String dataref

let dataref: Borrowed<String, ReadOnly>
     = Borrowed::find("sim/version/sim_build_string").unwrap();
let sim_build_time = dataref.get();

Writeable i32 dataref

let dataref: Borrowed<i32, ReadWrite>
    = Borrowed::find("sim/cockpit2/autopilot/flight_director_mode").unwrap();
let mode = dataref.get();
dataref.set(3);

Methods

impl<D, A> Borrowed<D, A> where D: DataType, A: DataAccess

fn find(name: &str) -> Result<Borrowed<D, A>, SearchError>

Finds a dataref with the provided name. Returns a Borrowed object or an error