Struct xplm::data::Owned

pub struct Owned<D, A> {
    // some fields omitted
}

A dataref created and owned by this plugin

The DataAccess type parameter for owned datarefs is the access that other plugins have. This plugin can always write datarefs that it owns.

Methods

impl<D, A> Owned<D, A> where D: DataType, A: DataAccess

fn create(name: &str, initial_value: D) -> Result<Owned<D, A>, NulError>

Creates a dataref with the provided name, set to the provided value

Returns the dataref on success, or an error if the provided name was invalid.