Trait xplm::data::ArrayWriteable

pub trait ArrayWriteable<T>: ArrayReadable<T> + Writeable<Vec<T>> {
    fn set_from_slice(&mut self, value: &[T]);
}

A trait for objects that can be written like arrays

Required Methods

fn set_from_slice(&mut self, value: &[T])

Sets the values in this array with the values from a slice.

If the slice has more than i32::max_value()elements, onlyi32::max_value()` elements will be set.

If the slice has more elements than this array, the extra values will be ignored.

If the slice has fewer elements than this array, the extra elements in this array will not be changed.