Struct xplm::features::Feature

pub struct Feature {
    // some fields omitted
}

A feature that can be enabled or disabled

Methods

impl Feature

fn with_name(name: &str) -> Option<Feature>

Creates a feature with a name

Returns None if X-Plane does not support the requested feature, or if the name contains one or more null bytes.

fn all() -> Vec<Feature>

Returns all the features that X-Plane supports

fn name(&self) -> Option<&str>

Returns the name of this feature

Returns None if the feature name is not a valid UTF-8 string.

fn set_enabled(&mut self, enabled: bool)

Enables or disables this feature