Struct xplm::flight_loop::FlightLoop

pub struct FlightLoop {
    // some fields omitted
}

A handle to something that calls a flight loop callback

A FlightLoop object stores a callback that will be called by X-Plane every time the flight model is calculated.

When a FlightLoop object goes out of scope, the callback will be unregistered from X-Plane and deleted.

Examples

A flight loop callback executed every time the flight model is calculated

let callback = || {
    println!("Flight loop callback running");
    NextCallback::after_loops(1)
};
let flight_loop = FlightLoop::new(Phase::AfterFlightModel, callback);
flight_loop.schedule(NextCallback::AfterLoops(1));

Methods

impl FlightLoop

fn new<C>(phase: Phase, callback: C) -> FlightLoop where C: 'static + FlightLoopCallback

Creates a flight loop callback that will execute the given callback in the given phase.

The callback will not be called until schedule() is called.

fn schedule(&self, time: NextCallback)

Schedules the callback to be executed