Module xplm::graphics

OpenGL-related functionality

Modules

window
Low-level windows

Structs

GraphicsState
Stores various flags that can be enabled or disabled

Functions

bind_texture
Binds a texture ID to a texture number

generate_texture_numbers
Generates texture numbers in a range not reserved for X-Plane.

set_state
Sets the graphics state