Struct xplm::graphics::window::Window

pub struct Window {
    // some fields omitted
}

A low-level window

A Window provides a designated area on the screen and callbacks for drawing and handling user input. By default, it is invisible and does not respond to input.

Windows should always be kept in Rc<RefCell>s to ensure proper functionality.

Methods

impl Window

fn new(geometry: &Rect) -> Rc<RefCell<Window>>

Creates a new window the the specified geometry

fn set_draw_callback<C>(&mut self, callback: C) where C: 'static + DrawCallback

Sets the callback used to draw this window

fn set_keyboard_callback<C>(&mut self, callback: C) where C: 'static + KeyboardCallback

Sets the callback to handle keyboard events for this window

fn set_mouse_callback<C>(&mut self, callback: C) where C: 'static + MouseCallback

Sets the callback to handle mouse events for this window

fn set_mouse_wheel_callback<C>(&mut self, callback: C) where C: 'static + MouseWheelCallback

Sets the callback to handle mouse wheel events for this window

fn set_cursor_callback<C>(&mut self, callback: C) where C: 'static + CursorCallback

Sets the callback to specify cursors for this window

fn set_visible(&mut self, visible: bool)

Sets a window to be visible or hidden

fn bring_to_front(&self)

Moves this window on top of other windows

fn request_focus(&self)

Gives this window keyboard focus

fn get_geometry(&self) -> Rect

Returns the geometry of this window

fn set_geometry(&mut self, geometry: &Rect)

Sets the geometry of this window