Module xplm::ipc

Inter-plugin communication Functionality for inter-plugin communication

Structs

Plugin
Represents a plugin and allows access to it

PluginInfo
Information about a plugin

SendError
An error that indicates that a message could not be sent because its message number was invald

Enums

XPlaneMessage
Messages that X-Plane can send to a plugin

Constants

MIN_USER_MESSAGE
Smallest message number allowed to be sent by a plugin (others are reserved for X-Plane)

XPLANE_ID
A plugin ID that indicates a message sent by X-Plane