Module xplm::nav

Provides access to the navigation database

Navigation database access

Structs

Airport
Represents an airport

DME
Represents a DME

Fix
Represents a fix

Glideslope
Represents an ILS glideslope

ILSLocalizer
Represents an ILS localizer (the glideslope is a separate Glideslope object)

InnerMarker
Represents an inner marker

Localizer
Represents a standalone localizer

MiddleMarker
Represents a middle marker

NDB
Represents a non-directional beacon

NavaidIterator
An iterator over available navaids

OuterMarker
Represents an outer marker

VOR
Represents a VOR

Enums

Navaid
Contains a navaid of any of the supported types

Functions

all_airports
Returns an iterator over all airports

all_dmes
Returns an iterator over all DMEs

all_fixes
Returns an iterator over all fixes

all_ils_glideslopes
Returns an iterator over all ILS glideslopes

all_ils_localizers
Returns an iterator over all ILS localizers

all_inner_markers
Returns an iterator over all inner markers

all_localizers
Returns an iterator over all standalone localizers

all_middle_markers
Returns an iterator over all middle markers

all_navaids
Returns an iterator over all available navaids in the database

all_ndbs
Returns an iterator over all NDBs

all_outer_markers
Returns an iterator over all outer markers

all_vors
Returns an iterator over all VORs