Trait xplm::Plugin

pub trait Plugin: Sized {
    fn start() -> Result<Self, Box<Error>>;
    fn enable(&mut self);
    fn disable(&mut self);
    fn info() -> PluginInfo;
    fn message_from_plugin(&mut self, message: i32, from: Plugin);
    fn message_from_xplane(&mut self, message: XPlaneMessage);
    fn stop(&mut self);
}

The trait that all plugins should implement

Required Methods

fn start() -> Result<Self, Box<Error>>

Called when X-Plane loads this plugin

On success, returns a plugin object.

On failure, returns an error.

fn enable(&mut self)

Called when the plugin is enabled

fn disable(&mut self)

Called when the plugin is disabled

fn info() -> PluginInfo

Returns information on this plugin

fn message_from_plugin(&mut self, message: i32, from: Plugin)

Called when a plugin sends a message to this plugin

The message number and the plugin that sent the message are provided. This method is not called when X-Plane sends a message.

fn message_from_xplane(&mut self, message: XPlaneMessage)

Called when X-Plane sends a message to this plugin

The message is provided.

fn stop(&mut self)

The plugin will be droped after this function is called.