Struct xplm::ui::widget::Pane

pub struct Pane {
    // some fields omitted
}

A pane that can subdivide a Window

While a Window automatically moves its children when the user drags it around the screen, a Pane does not. Therefore, it may be better to have UI elements be children of a window than children of a Pane.

Methods

impl Pane

fn new(title: &str, geometry: &Rect) -> Pane

Creates a pane with the provided title and geometry

fn set_pane_type(&mut self, pane_type: PaneType)

Sets the type of this pane