Struct xplm::ui::widget::Window

pub struct Window {
    // some fields omitted
}

Represents a window

By default, a window has close buttons.

The descriptor of a window is its title.

Methods

impl Window

fn new(title: &str, geometry: &Rect) -> Window

Creates a new Window with the provided title and geometry

fn set_close_buttons(&mut self, close_buttons: bool)

Sets whether this window should have close buttons

fn set_translucent(&mut self, translucent: bool)

Sets whether this window should be translucent or should be a standard opaque window