Trait xplm::ui::widget::Widget

pub trait Widget {
    fn widget_id(&self) -> XPWidgetID;
    fn set_visible(&mut self, visible: bool);
    fn get_property(&self, property: i32) -> Option<isize>;
    fn set_property(&mut self, property: i32, value: isize);
    fn get_descriptor(&self) -> String;
    fn set_descriptor(&mut self, descriptor: &str);
    fn get_geometry(&self) -> Rect;
    fn set_geometry(&mut self, geometry: &Rect);
    fn clear_children(&mut self);
    fn add_child(&mut self, child: Box<Widget>);
}

Common functions for all types of widgets

Required Methods

fn widget_id(&self) -> XPWidgetID

Returns the ID of this widget

fn set_visible(&mut self, visible: bool)

Shows or hides this widget

If this widget is not a root widget and is not a child of a root widget, showing it may not have any effect.

fn get_property(&self, property: i32) -> Option<isize>

Returns the value of a property, or None if this widget does not have the requested property

fn set_property(&mut self, property: i32, value: isize)

Sets the value of a property of this widget

fn get_descriptor(&self) -> String

Returns the descriptor of this widget

fn set_descriptor(&mut self, descriptor: &str)

Sets the descriptor of this widget

If the provided string is not valid as a C string, the descriptor will not be changed.

fn get_geometry(&self) -> Rect

Returns the geometry of this widget

fn set_geometry(&mut self, geometry: &Rect)

Sets the geometry of this widget

fn clear_children(&mut self)

Removes the children of this widget

fn add_child(&mut self, child: Box<Widget>)

Adds a child to this widget